Split main.py
