+++ models/__init__.py
@@ -1,3 +1,4 @@
 from .base_model import BaseModel
 from .fcn import FCN
 from .cnn import CNN
+from .lr import LinearRegression
