+++ models/fcn.py
@@ -2,6 +2,7 @@ import tensorflow as tf
 import logging
 import tf_variables
 from .base_model import BaseModel
+from .base_ops import *
 
 
 def two_layer(x, num_classes=10, dropout_keep_prob=0.5, hidden_layer=256):
@@ -25,18 +26,6 @@ def two_layer(x, num_classes=10, dropout_keep_prob=0.5, hidden_layer=256):
 
         y_conv = tf.matmul(h_fc1, W_fc2) + b_fc2
     return y_conv, None
-
-
-def weight_variable(shape):
-    """weight_variable generates a weight variable of a given shape."""
-    initial = tf.truncated_normal(shape, stddev=0.1)
-    return tf.Variable(initial)
-
-
-def bias_variable(shape):
-    """bias_variable generates a bias variable of a given shape."""
-    initial = tf.constant(0.1, shape=shape)
-    return tf.Variable(initial)
 
 
 # Two-Layer Fully-Connected Network
